Completed the refactoring of Ids. Used the info at https://www.ssw.com.au/rules/do-you-use-strongly-typed-ids/

+++ src/Application/Features/Artifacts/Queries/GetArtifactsQuery.cs
@@ -8,8 +8,8 @@ public record GetArtifactsQuery(
     int? Take = null,
     TextSearchType? SearchType = null,
     // Additional
-    IEnumerable<long>? VideoIds = null,
-    IEnumerable<long>? ArtifactIds = null,
+    IEnumerable<int>? VideoIds = null,
+    IEnumerable<int>? ArtifactIds = null,
     string? SearchOn = null) : IRequest<Page<ArtifactDTO>>, IBasicQuery;
 
 internal class GetArtifactsQueryValidator : AbstractValidator<GetArtifactsQuery>
